requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
google-genai>=0.2.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0
yt-dlp>=2024.12.0
moviepy>=1.0.3
aiofiles>=24.1.0
httpx>=0.27.0
python-multipart>=0.0.12
opencv-python>=4.8.0
numpy>=1.24.0
scipy>=1.11.0
pillow>=10.0.0

